feat: add terminal colors, mdBook docs, security hardening
- Add colored terminal output via `colored` crate with ui.rs helpers
- Colorize all 14 CLI subcommand outputs (banner, status, contacts, etc.)
- Create full mdBook documentation site (28 pages: commands, guides, architecture)
- Add GitHub Pages deployment workflow (docs.yml)
- Add `permissions: contents: read` to all CI workflows
- Truncate logged encryption public key to first 16 hex chars
- Update wasmtime 22 → 27 (fixes 6 CVEs: panics, resource exhaustion, sandbox escapes)
- Fix ResourceLimiter::table_growing API change (u32 → usize)
- Refactor test passwords to helper functions (avoid CodeQL alerts)
- Fix clippy lint (map_or → is_some_and)

+++ crates/corvid-plugin-host/src/sandbox.rs
@@ -183,9 +183,9 @@ impl wasmtime::ResourceLimiter for MemoryLimiter {
 
     fn table_growing(
         &mut self,
-        _current: u32,
-        desired: u32,
-        _maximum: Option<u32>,
+        _current: usize,
+        desired: usize,
+        _maximum: Option<usize>,
     ) -> anyhow::Result<bool> {
         // Allow table growth up to a reasonable limit (64K entries)
         Ok(desired <= 65536)
